+++ WaterSystem_v1/WaterSystem_Model/EmployeeModel.cs
@@ -8,13 +8,13 @@ namespace WaterSystem_Model
 {
     public class EmployeeModel: IModel
     {
-        public int Emp_Id { get; set; }
+        public decimal Emp_Id { get; set; }
         public string Emp_Name { get; set; }
         public string Emp_Add { get; set; }
-        public int Emp_Mob { get; set; }
+        public decimal Emp_Mob { get; set; }
         public DateTime Emp_BirthDate { get; set; }
         public DateTime Emp_JoinDate { get; set; }
-        public int Emp_Sal { get; set; }
+        public double Emp_Sal { get; set; }
         public bool IsCheck { get; set; }
         public string Update { get { return "Update"; } }
         public string Delete { get { return "Delete"; } }
